refactor: reorganized the styles folder structure
