+++ src/assets/styles/components/AboutSection.scss
@@ -1,5 +1,5 @@
-@import 'colours';
-@import 'breakpoints';
+@import '~assets/styles/common/colours';
+@import '~assets/styles/common/breakpoints';
 
 .about-section {
   padding: 2rem 4rem;
