+++ src/assets/styles/components/MainHeader.scss
@@ -1,5 +1,5 @@
-@import 'colours';
-@import 'breakpoints';
+@import '~assets/styles/common/colours';
+@import '~assets/styles/common/breakpoints';
 
 .main-header-container {
   background: $dark-gray;
