+++ src/assets/styles/components/ProjectCard.scss
@@ -1,5 +1,5 @@
-@import 'colours';
-@import 'mixins';
+@import '~assets/styles/common/colours';
+@import '~assets/styles/common/mixins';
 
 $short-transition-time: 0.7s;
