+++ src/assets/styles/components/ProjectsSection.scss
@@ -1,4 +1,4 @@
-@import 'colours';
+@import '~assets/styles/common/colours';
 
 .projects-section {
   padding: 2rem 4rem;
