+++ src/pages/index.tsx
@@ -3,7 +3,7 @@ import type { HeadFC, PageProps } from 'gatsby';
 import MainHeader from '../components/MainHeader';
 import ProjectsSection from '../components/ProjectsSection';
 import AboutSection from '../components/AboutSection';
-import 'assets/styles/global-styles.scss';
+import 'assets/styles/common/global-styles.scss';
 
 const IndexPage: React.FC<PageProps> = () => {
   return (
